output to and serve from dist

+++ src/project/clean.ts
@@ -3,5 +3,10 @@ import fs from 'fs-extra';
 import {paths} from '../paths.js';
 
 export const clean = async () => {
-	await fs.emptyDir(paths.build);
+	if (fs.existsSync(paths.build)) {
+		await fs.emptyDir(paths.build);
+	}
+	if (fs.existsSync(paths.dist)) {
+		await fs.emptyDir(paths.dist);
+	}
 };
